moved readers to sync module
