+++ src/readers/sync.rs
@@ -2,465 +2,7 @@
 
 use std::io::*;
 
-#[cfg(feature = "reader-asf")]
-mod asf {
-    use crate::FileFormat;
-
-    // Maximum number of descriptors that can be processed by the reader.
-    pub(super) const DESCRIPTOR_LIMIT: usize = 32;
-
-    // Maximum size of a descriptor name that can be handled by the reader.
-    pub(super) const DESCRIPTOR_NAME_LIMIT: usize = 64;
-
-    // Maximum number of objects that can be processed by the reader.
-    pub(super) const OBJECT_LIMIT: usize = 256;
-
-    // GUID for extended content description object.
-    pub(super) const EXTENDED_CONTENT_DESCRIPTION_OBJECT_GUID: &str =
-        "d2d0a440-e307-11d2-97f0-00a0c95ea850";
-
-    // GUID for stream properties object.
-    pub(super) const STREAM_PROPERTIES_OBJECT_GUID: &str = "b7dc0791-a9b7-11cf-8ee6-00c00c205365";
-
-    // GUID for audio media.
-    pub(super) const AUDIO_MEDIA_GUID: &str = "f8699e40-5b4d-11cf-a8fd-00805f5c442b";
-
-    // GUID for video media.
-    pub(super) const VIDEO_MEDIA_GUID: &str = "bc19efc0-5b4d-11cf-a8fd-00805f5c442b";
-
-    // Determines the file format based on the identified streams.
-    pub(super) fn determine_file_format(video_stream: bool, audio_stream: bool) -> FileFormat {
-        use FileFormat as F;
-        if video_stream {
-            F::WindowsMediaVideo
-        } else if audio_stream {
-            F::WindowsMediaAudio
-        } else {
-            F::AdvancedSystemsFormat
-        }
-    }
-}
-
-#[cfg(feature = "reader-cfb")]
-mod cfb {
-    use crate::FileFormat;
-
-    // Determines the file format based on the CLSID.
-    pub(super) fn determine_file_format(clsid: &str) -> Option<FileFormat> {
-        use FileFormat as F;
-        match clsid {
-            "e60f81e1-49b3-11d0-93c3-7e0706000000" => Some(F::AutodeskInventorAssembly),
-            "bbf9fdf1-52dc-11d0-8c04-0800090be8ec" => Some(F::AutodeskInventorDrawing),
-            "4d29b490-49b2-11d0-93c3-7e0706000000" => Some(F::AutodeskInventorPart),
-            "76283a80-50dd-11d3-a7e3-00c04f79d7bc" => Some(F::AutodeskInventorPresentation),
-            "402efe62-1999-101b-99ae-04021c007002" => Some(F::CorelPresentations7),
-            "597caa70-72aa-11cf-831e-524153480000" => Some(F::FlashProject),
-            "b4f235fe-dca6-4803-b7b2-c25a453c836b" => Some(F::FlashProject),
-            "00020810-0000-0000-c000-000000000046" => Some(F::MicrosoftExcelSpreadsheet),
-            "00020820-0000-0000-c000-000000000046" => Some(F::MicrosoftExcelSpreadsheet),
-            "00044851-0000-0000-c000-000000000046" => Some(F::MicrosoftPowerpointPresentation),
-            "64818d10-4f9b-11cf-86ea-00aa00b929e8" => Some(F::MicrosoftPowerpointPresentation),
-            "ea7bae70-fb3b-11cd-a903-00aa00510ea3" => Some(F::MicrosoftPowerpointPresentation),
-            "74b78f3a-c8c8-11d1-be11-00c04fb6faf1" => Some(F::MicrosoftProjectPlan),
-            "00021201-0000-0000-00c0-000000000046" => Some(F::MicrosoftPublisherDocument),
-            "000c1084-0000-0000-c000-000000000046" => Some(F::MicrosoftSoftwareInstaller),
-            "00021a13-0000-0000-c000-000000000046" => Some(F::MicrosoftVisioDrawing),
-            "00021a14-0000-0000-c000-000000000046" => Some(F::MicrosoftVisioDrawing),
-            "00020900-0000-0000-c000-000000000046" => Some(F::MicrosoftWordDocument),
-            "00020906-0000-0000-c000-000000000046" => Some(F::MicrosoftWordDocument),
-            "00021303-0000-0000-c000-000000000046" => Some(F::MicrosoftWorksDatabase),
-            "28cddbc3-0ae2-11ce-a29a-00aa004a1a72" => Some(F::MicrosoftWorksDatabase),
-            "00021302-0000-0000-c000-000000000046" => Some(F::MicrosoftWorksWordProcessor),
-            "0ea45ab2-9e0a-11d1-a407-00c04fb932ba" => Some(F::MicrosoftWorksWordProcessor),
-            "28cddbc2-0ae2-11ce-a29a-00aa004a1a72" => Some(F::MicrosoftWorksWordProcessor),
-            "83a33d36-27c5-11ce-bfd4-00400513bb57" => Some(F::SolidworksAssembly),
-            "83a33d34-27c5-11ce-bfd4-00400513bb57" => Some(F::SolidworksDrawing),
-            "83a33d30-27c5-11ce-bfd4-00400513bb57" => Some(F::SolidworksPart),
-            "3f543fa0-b6a6-101b-9961-04021c007002" => Some(F::Starcalc),
-            "6361d441-4235-11d0-89cb-008029e4b0b1" => Some(F::Starcalc),
-            "c6a5b861-85d6-11d1-89cb-008029e4b0b1" => Some(F::Starcalc),
-            "02b3b7e0-4225-11d0-89ca-008029e4b0b1" => Some(F::Starchart),
-            "bf884321-85dd-11d1-89d0-008029e4b0b1" => Some(F::Starchart),
-            "fb9c99e0-2c6d-101c-8e2c-00001b4cc711" => Some(F::Starchart),
-            "2e8905a0-85bd-11d1-89d0-008029e4b0b1" => Some(F::Stardraw),
-            "af10aae0-b36d-101b-9961-04021c007002" => Some(F::Stardraw),
-            "012d3cc0-4216-11d0-89cb-008029e4b0b1" => Some(F::Starimpress),
-            "565c7221-85bc-11d1-89d0-008029e4b0b1" => Some(F::Starimpress),
-            "02b3b7e1-4225-11d0-89ca-008029e4b0b1" => Some(F::Starmath),
-            "d4590460-35fd-101c-b12a-04021c007002" => Some(F::Starmath),
-            "ffb5e640-85de-11d1-89d0-008029e4b0b1" => Some(F::Starmath),
-            "8b04e9b0-420e-11d0-a45e-00a0249d57b1" => Some(F::Starwriter),
-            "c20cf9d1-85ae-11d1-aab4-006097da561a" => Some(F::Starwriter),
-            "dc5c7e40-b35c-101b-9961-04021c007002" => Some(F::Starwriter),
-            "1cdd8c7b-81c0-45a0-9fed-04143144cc1e" => Some(F::ThreeDimensionalStudioMax),
-            "519873ff-2dad-0220-1937-0000929679cd" => Some(F::WordperfectDocument),
-            "402efe60-1999-101b-99ae-04021c007002" => Some(F::WordperfectGraphics),
-            _ => None,
-        }
-    }
-}
-
-#[cfg(feature = "reader-ebml")]
-mod ebml {
-    use crate::FileFormat;
-
-    // Maximum number of EBML elements that can be processed by the reader.
-    pub(super) const ELEMENT_LIMIT: usize = 256;
-
-    // Maximum size of a Codec ID that can be processed by the reader.
-    pub(super) const CODEC_ID_LIMIT: usize = 64;
-
-    // Maximum size of a DocType that can be processed by the reader.
-    pub(super) const DOC_TYPE_LIMIT: usize = 8;
-
-    // DocType element ID.
-    pub(super) const DOC_TYPE_ELEMENT_ID: u32 = 0x4282;
-
-    // EBML element ID.
-    pub(super) const EBML_ELEMENT_ID: u32 = 0x1A45DFA3;
-
-    // Cluster element ID.
-    pub(super) const CLUSTER_ELEMENT_ID: u32 = 0x1F43B675;
-
-    // CodecID element ID.
-    pub(super) const CODEC_ID_ELEMENT_ID: u32 = 0x86;
-
-    // Segment element ID.
-    pub(super) const SEGMENT_ELEMENT_ID: u32 = 0x18538067;
-
-    // StereoMode element ID.
-    pub(super) const STEREO_MODE_ELEMENT_ID: u32 = 0x53B8;
-
-    // Tracks element ID.
-    pub(super) const TRACKS_ELEMENT_ID: u32 = 0x1654AE6B;
-
-    // TrackEntry element ID.
-    pub(super) const TRACK_ENTRY_ELEMENT_ID: u32 = 0xAE;
-
-    // Video element ID.
-    pub(super) const VIDEO_ELEMENT_ID: u32 = 0xE0;
-
-    // Determines the file format based on the identified tracks.
-    pub(super) fn determine_file_format(
-        video_track: bool,
-        audio_track: bool,
-        subtitle_track: bool,
-    ) -> FileFormat {
-        use FileFormat as F;
-        if video_track {
-            F::MatroskaVideo
-        } else if audio_track {
-            F::MatroskaAudio
-        } else if subtitle_track {
-            F::MatroskaSubtitles
-        } else {
-            F::ExtensibleBinaryMetaLanguage
-        }
-    }
-}
-
-#[cfg(feature = "reader-mp4")]
-mod mp4 {
-    use crate::FileFormat;
-
-    // Maximum number of boxes that can be processed by the reader.
-    pub(super) const BOX_LIMIT: usize = 256;
-
-    // Determines the file format based on the identified tracks.
-    pub(super) fn determine_file_format(
-        video_track: bool,
-        audio_track: bool,
-        subtitle_track: bool,
-    ) -> FileFormat {
-        use FileFormat as F;
-        if video_track {
-            F::Mpeg4Part14Video
-        } else if audio_track {
-            F::Mpeg4Part14Audio
-        } else if subtitle_track {
-            F::Mpeg4Part14Subtitles
-        } else {
-            F::Mpeg4Part14
-        }
-    }
-}
-
-#[cfg(feature = "reader-pdf")]
-mod pdf {
-    // Maximum number of bytes that can be processed by the reader (32 MB).
-    pub(super) const READ_LIMIT: usize = 33_554_432;
-
-    // Size of each chunk to read (32 KB).
-    pub(super) const CHUNK_SIZE: usize = 32_768;
-
-    // Size of overlap to keep between chunks.
-    pub(super) const OVERLAP_SIZE: usize = AI_MARKER.len() - 1;
-
-    // Marker for the AI file format.
-    pub(super) const AI_MARKER: &[u8] = b"AIPrivateData";
-}
-
-#[cfg(feature = "reader-rm")]
-mod rm {
-    use crate::FileFormat;
-
-    // Maximum number of chunks that can be processed by the reader.
-    pub(super) const CHUNK_LIMIT: usize = 64;
-
-    // Determines the file format based on the identified streams.
-    pub(super) fn determine_file_format(video_stream: bool, audio_stream: bool) -> FileFormat {
-        use FileFormat as F;
-        if video_stream {
-            F::Realvideo
-        } else if audio_stream {
-            F::Realaudio
-        } else {
-            F::Realmedia
-        }
-    }
-}
-
-#[cfg(feature = "reader-sqlite3")]
-mod sqlite3 {
-    use crate::FileFormat;
-
-    use super::FindBytes;
-
-    // Marker for the Sketch file format.
-    const SKETCH_MARKER: &[u8] = b"com.bohemiancoding.sketch3";
-
-    // Checks if the buffer holds the Sketch file format marker.
-    pub(super) fn check_if_buffer_holds_sketch_marker(
-        buf: &[u8],
-        nread: usize,
-    ) -> Option<FileFormat> {
-        if buf[..nread].holds(SKETCH_MARKER) {
-            Some(FileFormat::Sketch)
-        } else {
-            None
-        }
-    }
-}
-
-#[cfg(feature = "reader-txt")]
-mod txt {
-    use std::io::{Error, ErrorKind};
-
-    // Maximum number of lines that can be processed by the reader.
-    pub(super) const LINE_LIMIT: usize = 16;
-
-    // Maximum number of bytes that can be processed by the reader (64 KB).
-    pub(super) const READ_LIMIT: u64 = 65_536;
-
-    // Checks for control characters other than whitespaces.
-    pub(super) fn check_for_non_whitespace_control_char(
-        line: Result<String, Error>,
-    ) -> Result<(), Error> {
-        line?
-            .chars()
-            .find(|char| char.is_control() && !char.is_whitespace())
-            .map(|_| Err(Error::new(ErrorKind::InvalidData, "invalid characters")))
-            .unwrap_or(Ok(()))
-    }
-}
-
-#[cfg(feature = "reader-xml")]
-mod xml {
-    use crate::{readers::FindBytes, FileFormat};
-
-    // Checks if the buffer holds markers indicating the presence of various file formats.
-    pub(super) fn check_if_buffer_holds_markers(buf: &[u8]) -> FileFormat {
-        use FileFormat as F;
-        if buf.holds("<abiword template=\"false\"") {
-            F::Abiword
-        } else if buf.holds("<abiword template=\"true\"") {
-            F::AbiwordTemplate
-        } else if buf.holds("<amf") {
-            F::AdditiveManufacturingFormat
-        } else if buf.holds("<ASX") || buf.holds("<asx") {
-            F::AdvancedStreamRedirector
-        } else if buf.holds("<feed") {
-            F::Atom
-        } else if buf.holds("<COLLADA") {
-            F::CollaborativeDesignActivity
-        } else if buf.holds("<mxfile") {
-            F::Drawio
-        } else if buf.holds("<X3D") {
-            F::Extensible3d
-        } else if buf.holds("<xsl") {
-            F::ExtensibleStylesheetLanguageTransformations
-        } else if buf.holds("<FictionBook") {
-            F::Fictionbook
-        } else if buf.holds("<gml") {
-            F::GeographyMarkupLanguage
-        } else if buf.holds("<gpx") {
-            F::GpsExchangeFormat
-        } else if buf.holds("<kml") {
-            F::KeyholeMarkupLanguage
-        } else if buf.holds("<math") {
-            F::MathematicalMarkupLanguage
-        } else if buf.holds("<MPD") {
-            F::MpegDashMpd
-        } else if buf.holds("<score-partwise") {
-            F::Musicxml
-        } else if buf.holds("<rss") {
-            F::ReallySimpleSyndication
-        } else if buf.holds("<SVG") || buf.holds("<svg") {
-            F::ScalableVectorGraphics
-        } else if buf.holds("<soap") {
-            F::SimpleObjectAccessProtocol
-        } else if buf.holds("<map") {
-            F::TiledMapXml
-        } else if buf.holds("<tileset") {
-            F::TiledTilesetXml
-        } else if buf.holds("<tt") && buf.holds("xmlns=\"http://www.w3.org/ns/ttml\"") {
-            F::TimedTextMarkupLanguage
-        } else if buf.holds("<TrainingCenterDatabase") {
-            F::TrainingCenterXml
-        } else if buf.holds("<uof:UOF") && buf.holds("uof:mimetype=\"vnd.uof.presentation\"") {
-            F::UniformOfficeFormatPresentation
-        } else if buf.holds("<uof:UOF") && buf.holds("uof:mimetype=\"vnd.uof.spreadsheet\"") {
-            F::UniformOfficeFormatSpreadsheet
-        } else if buf.holds("<uof:UOF") && buf.holds("uof:mimetype=\"vnd.uof.text\"") {
-            F::UniformOfficeFormatText
-        } else if buf.holds("<USFSubtitles") {
-            F::UniversalSubtitleFormat
-        } else if buf.holds("<xliff") {
-            F::XmlLocalizationInterchangeFileFormat
-        } else if buf.holds("<playlist") {
-            F::XmlShareablePlaylistFormat
-        } else {
-            F::ExtensibleMarkupLanguage
-        }
-    }
-}
-
-#[cfg(feature = "reader-zip")]
-mod zip {
-    use crate::FileFormat;
-
-    // Maximum number of entries that can be processed by the reader.
-    pub(super) const ENTRY_LIMIT: usize = 1024;
-
-    // Signature of the ZIP64 end of central directory locator.
-    pub(super) const EOCD64_LOCATOR_SIGNATURE: &[u8] = b"PK\x06\x07";
-
-    // Signature of the end of central directory record.
-    pub(super) const EOCD_SIGNATURE: &[u8] = b"PK\x05\x06";
-
-    // Size of the ZIP64 end of central directory locator.
-    pub(super) const EOCD64_LOCATOR_SIZE: usize = 20;
-
-    // Maximum size of the end of central directory record.
-    pub(super) const EOCD_MAX_SIZE: usize = EOCD_MIN_SIZE + u16::MAX as usize;
-
-    // Minimum size of the end of central directory record.
-    pub(super) const EOCD_MIN_SIZE: usize = 22;
-
-    // Checks the trimmed data.
-    pub(super) fn check_trimmed_data(data: String) -> FileFormat {
-        use FileFormat as F;
-        match data.trim() {
-            "application/epub+zip" => F::ElectronicPublication,
-            "application/vnd.adobe.indesign-idml-package" => F::IndesignMarkupLanguage,
-            "application/vnd.oasis.opendocument.base"
-            | "application/vnd.oasis.opendocument.database" => F::OpendocumentDatabase,
-            "application/vnd.oasis.opendocument.formula" => F::OpendocumentFormula,
-            "application/vnd.oasis.opendocument.formula-template" => F::OpendocumentFormulaTemplate,
-            "application/vnd.oasis.opendocument.graphics" => F::OpendocumentGraphics,
-            "application/vnd.oasis.opendocument.graphics-template" => {
-                F::OpendocumentGraphicsTemplate
-            }
-            "application/vnd.oasis.opendocument.presentation" => F::OpendocumentPresentation,
-            "application/vnd.oasis.opendocument.presentation-template" => {
-                F::OpendocumentPresentationTemplate
-            }
-            "application/vnd.oasis.opendocument.spreadsheet" => F::OpendocumentSpreadsheet,
-            "application/vnd.oasis.opendocument.spreadsheet-template" => {
-                F::OpendocumentSpreadsheetTemplate
-            }
-            "application/vnd.oasis.opendocument.text" => F::OpendocumentText,
-            "application/vnd.oasis.opendocument.text-master" => F::OpendocumentTextMaster,
-            "application/vnd.oasis.opendocument.text-master-template" => {
-                F::OpendocumentTextMasterTemplate
-            }
-            "application/vnd.oasis.opendocument.text-template" => F::OpendocumentTextTemplate,
-            "application/vnd.recordare.musicxml" => F::MusicxmlZip,
-            "application/vnd.sun.xml.calc" => F::SunXmlCalc,
-            "application/vnd.sun.xml.calc.template" => F::SunXmlCalcTemplate,
-            "application/vnd.sun.xml.draw" => F::SunXmlDraw,
-            "application/vnd.sun.xml.draw.template" => F::SunXmlDrawTemplate,
-            "application/vnd.sun.xml.impress" => F::SunXmlImpress,
-            "application/vnd.sun.xml.impress.template" => F::SunXmlImpressTemplate,
-            "application/vnd.sun.xml.math" => F::SunXmlMath,
-            "application/vnd.sun.xml.writer" => F::SunXmlWriter,
-            "application/vnd.sun.xml.writer.global" => F::SunXmlWriterGlobal,
-            "application/vnd.sun.xml.writer.template" => F::SunXmlWriterTemplate,
-            "image/openraster" => F::Openraster,
-            _ => F::Zip,
-        }
-    }
-
-    // Checks the filename.
-    pub(crate) fn check_filename(filename: &str) -> Option<FileFormat> {
-        use FileFormat as F;
-        match filename {
-            "AndroidManifest.xml" => Some(F::AndroidPackage),
-            "AppManifest.xaml" => Some(F::Xap),
-            "AppxManifest.xml" => Some(F::WindowsAppPackage),
-            "AppxMetadata/AppxBundleManifest.xml" => Some(F::WindowsAppBundle),
-            "BundleConfig.pb" => Some(F::AndroidAppBundle),
-            "DOMDocument.xml" => Some(F::FlashCs5Project),
-            "META-INF/AIR/application.xml" => Some(F::AdobeIntegratedRuntime),
-            "META-INF/application.xml" => Some(F::EnterpriseApplicationArchive),
-            "META-INF/mozilla.rsa" => Some(F::Xpinstall),
-            "WEB-INF/web.xml" => Some(F::WebApplicationArchive),
-            "doc.kml" => Some(F::KeyholeMarkupLanguageZip),
-            "document.json" => Some(F::Sketch43),
-            "extension.vsixmanifest" => Some(F::MicrosoftVisualStudioExtension),
-            _ => None,
-        }
-    }
-
-    pub(crate) fn check_filename_starts_with(filename: String) -> Option<FileFormat> {
-        use FileFormat as F;
-        if filename.starts_with("Fusion[Active]/") {
-            Some(F::Autodesk123d)
-        } else if filename.starts_with("circuitdiagram/") {
-            Some(F::CircuitDiagramDocument)
-        } else if filename.starts_with("dwf/") {
-            Some(F::DesignWebFormatXps)
-        } else if filename.ends_with(".fb2") && !filename.contains('/') {
-            Some(F::FictionbookZip)
-        } else if filename.starts_with("FusionAssetName[Active]/") {
-            Some(F::Fusion360)
-        } else if filename.starts_with("Payload/") && filename.contains(".app/") {
-            Some(F::IosAppStorePackage)
-        } else if filename.starts_with("word/") {
-            Some(F::OfficeOpenXmlDocument)
-        } else if filename.starts_with("visio/") {
-            Some(F::OfficeOpenXmlDrawing)
-        } else if filename.starts_with("ppt/") {
-            Some(F::OfficeOpenXmlPresentation)
-        } else if filename.starts_with("xl/") {
-            Some(F::OfficeOpenXmlSpreadsheet)
-        } else if filename.starts_with("Documents/") && filename.ends_with(".fpage") {
-            Some(F::Openxps)
-        } else if filename.starts_with("SpaceClaim/") {
-            Some(F::SpaceclaimDocument)
-        } else if filename.starts_with("3D/") && filename.ends_with(".model") {
-            Some(F::ThreeDimensionalManufacturingFormat)
-        } else if (filename.ends_with(".usd")
-            || filename.ends_with(".usda")
-            || filename.ends_with(".usdc"))
-            && !filename.contains('/')
-        {
-            Some(F::UniversalSceneDescriptionZip)
-        } else {
-            None
-        }
-    }
-}
+use super::common::FindBytes;
 
 impl crate::FileFormat {
     /// Determines file format from the specified format reader, if any.
@@ -513,8 +55,8 @@ impl crate::FileFormat {
 
     /// Determines file format from an ASF reader.
     #[cfg(feature = "reader-asf")]
-    pub(crate) fn from_asf_reader<R: Read + Seek>(reader: R) -> Result<Self> {
-        use asf::{
+    fn from_asf_reader<R: Read + Seek>(reader: R) -> Result<Self> {
+        use super::common::asf::{
             determine_file_format, AUDIO_MEDIA_GUID, DESCRIPTOR_LIMIT, DESCRIPTOR_NAME_LIMIT,
             EXTENDED_CONTENT_DESCRIPTION_OBJECT_GUID, OBJECT_LIMIT, STREAM_PROPERTIES_OBJECT_GUID,
             VIDEO_MEDIA_GUID,
@@ -606,10 +148,10 @@ impl crate::FileFormat {
 
     /// Determines file format from a CFB reader.
     #[cfg(feature = "reader-cfb")]
-    pub(crate) fn from_cfb_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
+    fn from_cfb_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
         // Reads the major version.
 
-        use cfb::determine_file_format;
+        use super::common::cfb::determine_file_format;
         reader.seek(SeekFrom::Start(26))?;
         let major_version = reader.read_u16_le()?;
 
@@ -656,8 +198,8 @@ impl crate::FileFormat {
 
     /// Determines file format from an EBML reader.
     #[cfg(feature = "reader-ebml")]
-    pub(crate) fn from_ebml_reader<R: Read + Seek>(reader: R) -> Result<Self> {
-        use ebml::{
+    fn from_ebml_reader<R: Read + Seek>(reader: R) -> Result<Self> {
+        use super::common::ebml::{
             determine_file_format, CLUSTER_ELEMENT_ID, CODEC_ID_ELEMENT_ID, CODEC_ID_LIMIT,
             DOC_TYPE_ELEMENT_ID, DOC_TYPE_LIMIT, EBML_ELEMENT_ID, ELEMENT_LIMIT,
             SEGMENT_ELEMENT_ID, STEREO_MODE_ELEMENT_ID, TRACKS_ELEMENT_ID, TRACK_ENTRY_ELEMENT_ID,
@@ -784,7 +326,7 @@ impl crate::FileFormat {
 
     /// Determines file format from an EXE reader.
     #[cfg(feature = "reader-exe")]
-    pub(crate) fn from_exe_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
+    fn from_exe_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
         // Retrieves the stream length.
         let len = reader.seek(SeekFrom::End(0))?;
 
@@ -892,8 +434,8 @@ impl crate::FileFormat {
 
     /// Determines file format from a MP4 reader.
     #[cfg(feature = "reader-mp4")]
-    pub(crate) fn from_mp4_reader<R: Read + Seek>(reader: R) -> Result<Self> {
-        use mp4::{determine_file_format, BOX_LIMIT};
+    fn from_mp4_reader<R: Read + Seek>(reader: R) -> Result<Self> {
+        use super::common::mp4::{determine_file_format, BOX_LIMIT};
 
         // Creates a buffered reader.
         let mut reader = BufReader::new(reader);
@@ -964,8 +506,8 @@ impl crate::FileFormat {
 
     /// Determines file format from a PDF reader.
     #[cfg(feature = "reader-pdf")]
-    pub(crate) fn from_pdf_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
-        use pdf::{AI_MARKER, CHUNK_SIZE, OVERLAP_SIZE, READ_LIMIT};
+    fn from_pdf_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
+        use super::common::pdf::{AI_MARKER, CHUNK_SIZE, OVERLAP_SIZE, READ_LIMIT};
 
         // Rewinds to the beginning of the stream plus the size of the PDF file format signature.
         reader.seek(SeekFrom::Start(5))?;
@@ -1003,8 +545,8 @@ impl crate::FileFormat {
 
     /// Determines file format from a RM reader.
     #[cfg(feature = "reader-rm")]
-    pub(crate) fn from_rm_reader<R: Read + Seek>(reader: R) -> Result<Self> {
-        use rm::{determine_file_format, CHUNK_LIMIT};
+    fn from_rm_reader<R: Read + Seek>(reader: R) -> Result<Self> {
+        use super::common::rm::{determine_file_format, CHUNK_LIMIT};
 
         // Creates a buffered reader.
         let mut reader = BufReader::new(reader);
@@ -1064,8 +606,8 @@ impl crate::FileFormat {
 
     /// Determines file format from a SQLite 3 reader.
     #[cfg(feature = "reader-sqlite3")]
-    pub(crate) fn from_sqlite3_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
-        use sqlite3::check_if_buffer_holds_sketch_marker;
+    fn from_sqlite3_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
+        use super::common::sqlite3::check_if_buffer_holds_sketch_marker;
 
         // Rewinds to the beginning of the stream plus the size of the SQLite 3 file format header.
         reader.seek(SeekFrom::Start(100))?;
@@ -1085,8 +627,8 @@ impl crate::FileFormat {
 
     /// Determines file format from a TXT reader.
     #[cfg(feature = "reader-txt")]
-    pub(crate) fn from_txt_reader<R: Read + Seek>(reader: R) -> Result<Self> {
-        use txt::{check_for_non_whitespace_control_char, LINE_LIMIT, READ_LIMIT};
+    fn from_txt_reader<R: Read + Seek>(reader: R) -> Result<Self> {
+        use super::common::txt::{check_for_non_whitespace_control_char, LINE_LIMIT, READ_LIMIT};
 
         // Creates a buffered reader.
         let mut reader = BufReader::new(reader);
@@ -1106,10 +648,10 @@ impl crate::FileFormat {
 
     /// Determines file format from a XML reader.
     #[cfg(feature = "reader-xml")]
-    pub(crate) fn from_xml_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
-        // Rewinds to the beginning of the stream plus the size of the XML file format signature.
+    fn from_xml_reader<R: Read + Seek>(mut reader: R) -> Result<Self> {
+        use super::common::xml::check_if_buffer_holds_markers;
 
-        use xml::check_if_buffer_holds_markers;
+        // Rewinds to the beginning of the stream plus the size of the XML file format signature.
         reader.seek(SeekFrom::Start(5))?;
 
         // Creates and fills a buffer.
@@ -1123,8 +665,8 @@ impl crate::FileFormat {
 
     /// Determines file format from a ZIP reader.
     #[cfg(feature = "reader-zip")]
-    pub(crate) fn from_zip_reader<R: Read + Seek>(reader: R) -> Result<Self> {
-        use zip::{
+    fn from_zip_reader<R: Read + Seek>(reader: R) -> Result<Self> {
+        use super::common::zip::{
             check_filename, check_filename_starts_with, check_trimmed_data, ENTRY_LIMIT,
             EOCD64_LOCATOR_SIGNATURE, EOCD64_LOCATOR_SIZE, EOCD_MAX_SIZE, EOCD_MIN_SIZE,
             EOCD_SIGNATURE,
@@ -1349,92 +891,3 @@ trait ReadData: Read {
 /// Allows any type `R` that implements the `Read` trait to automatically benefit from the
 /// additional methods provided by the `ReadData` trait.
 impl<R: Read> ReadData for R {}
-
-/// A trait for finding a byte pattern within data.
-#[allow(dead_code)]
-trait FindBytes: AsRef<[u8]> {
-    /// Searches for the specified byte pattern and returns the index of the first occurrence.
-    fn find<P: AsRef<[u8]>>(&self, pat: P) -> Option<usize> {
-        // Retrieves references to data and pattern.
-        let data = self.as_ref();
-        let pat = pat.as_ref();
-
-        // An empty pattern is always considered to be contained in the data.
-        if pat.is_empty() {
-            return Some(0);
-        }
-
-        // The data is shorter than the pattern, so it cannot contain it.
-        if data.len() < pat.len() {
-            return None;
-        }
-
-        // Searches for the byte pattern using a modified Boyer-Moore-Horspool algorithm for forward
-        // search.
-        let mut shift_table = [pat.len(); 256];
-        for (index, &byte) in pat.iter().enumerate().take(pat.len() - 1) {
-            shift_table[byte as usize] = pat.len() - 1 - index;
-        }
-        let mut data_index = pat.len() - 1;
-        while data_index < data.len() {
-            let mut pat_index = pat.len() - 1;
-            while pat[pat_index] == data[data_index - (pat.len() - 1 - pat_index)] {
-                if pat_index == 0 {
-                    return Some(data_index - (pat.len() - 1));
-                }
-                pat_index -= 1;
-            }
-            data_index += shift_table[data[data_index] as usize];
-        }
-        None
-    }
-
-    /// Returns `true` if the data holds the specified byte pattern.
-    #[inline]
-    fn holds<P: AsRef<[u8]>>(&self, pat: P) -> bool {
-        self.find(pat).is_some()
-    }
-
-    /// Searches for the specified byte pattern and returns the index of the last occurrence.
-    fn rfind<P: AsRef<[u8]>>(&self, pat: P) -> Option<usize> {
-        // Retrieves references to data and pattern.
-        let data = self.as_ref();
-        let pat = pat.as_ref();
-
-        // An empty pattern is always considered to be contained in the data.
-        if pat.is_empty() {
-            return Some(data.len());
-        }
-
-        // The data is shorter than the pattern, so it cannot contain it.
-        if data.len() < pat.len() {
-            return None;
-        }
-
-        // Searches for the byte pattern using a modified Boyer-Moore-Horspool algorithm for reverse
-        // search.
-        let mut shift_table = [pat.len(); 256];
-        for (index, &byte) in pat.iter().rev().enumerate().take(pat.len() - 1) {
-            shift_table[byte as usize] = pat.len() - 1 - index;
-        }
-        let mut data_index = data.len() - pat.len();
-        loop {
-            let mut pat_index = 0;
-            while pat[pat_index] == data[data_index + pat_index] {
-                if pat_index == pat.len() - 1 {
-                    return Some(data_index);
-                }
-                pat_index += 1;
-            }
-            data_index = match data_index.checked_sub(shift_table[data[data_index] as usize]) {
-                Some(data_index) => data_index,
-                _ => break,
-            }
-        }
-        None
-    }
-}
-
-/// Allows any type `B` that implements the `AsRef<[u8]>` trait to benefit from the additional
-/// methods provided by the `FindBytes` trait.
-impl<B: AsRef<[u8]> + ?Sized> FindBytes for B {}
